+++ shardai-compiler/src/compiler.rs
@@ -35,7 +35,7 @@ impl Compiler {
             signature: *b"SBC",
             version_major: VERSION_MAJOR,
             version_minor: VERSION_MINOR,
-            constant_count: self.constants.len() as u32
+            constant_count: self.constants.len() as u16
         }
     }
